// src/plugins/storage-sqlite/sqlite-types.ts
export type SQLResultRow = {
    id: string;
    /**
     * Because we store the document fields as json,
     * just return a string here.
     */
    data: string;
} | [string, string, number, number, string];

export type SQLiteBasics<SQLiteDatabaseType> = {
    debugId?: string;

    /**
     * Opens a new database connection
     */
    open: (name: string) => Promise<SQLiteDatabaseType>;
    /**
     * Returns the query result rows
     */
    all(
        db: SQLiteDatabaseType,
        queryWithParams: SQLiteQueryWithParams
    ): Promise<SQLResultRow[]>;
    /**
     * Run a query. Return nothing.
     */
    run(
        db: SQLiteDatabaseType,
        queryWithParams: SQLiteQueryWithParams
    ): Promise<void>;
    /**
     * Sets a pragma like the WAL mode or other flags.
     * This cannot be done in run() or all() because
     * it does not return rows on some SQLite libraries
     * which would throw an error.
     * @link https://www.sqlite.org/pragma.html
     */
    setPragma(
        db: SQLiteDatabaseType,
        key: string,
        value: string
    ): Promise<void>;
    // closes the database connection
    close(db: SQLiteDatabaseType): Promise<void>;

    /**
     * [default=WAL2]
     * If empty string is given, the journalMode will be left untouched.
     * For example android has WAL as default, so we do not want to touch that setting.
     */
    journalMode: 'WAL' | 'WAL2' | 'DELETE' | 'TRUNCATE' | 'PERSIST' | 'MEMORY' | 'OFF' | '';
}
// ...
export type SQLiteQueryWithParams = {
    query: string;
    /**
     * Some SQLite version allow to use named params like $docId
     * and then putting an object in here.
     * But because not all environments support that,
     * we have to use plain array params that use the '?' placeholder in the query string.
     */
    params: (string | number | boolean)[];

    /**
     * Context must be set, will be used during debugging and error logs
     * to better show what is going wrong.
     */
    context: {
        method: string;
        data: any;
    };
}